Merged PR 14739: TASK-28058: add extra ListTile properties
These are required to allow us to move the Recent Documents resolver from Chang-e to the Realtime Svc.

This work is part of the transition process from the Throttler Svc to the ThrottlerNG Svc. Rather than update Chang-e so it can use the new service, I've decided that it's easier and more in line with out plans to decomission Chang-e to move the the Recent Documents resolver to the Realtime Svc which can already connect to the new service.

Related work items: #28058